ottimizzazione migliore fino ad ora,
need test se e abbastanza!

+++ model/Pin.h
@@ -8,11 +8,11 @@ class Pin {
 public:
     uint8_t number;      // Usa uint8_t per ridurre la dimensione
     PinType type;
-    float voltage;
+    uint16_t voltage;
     uint8_t isInput;     // Usa uint8_t per risparmiare spazio
     char note[50];       // Usa array di char per gestire le note
 
-    Pin(uint8_t num, PinType t, uint8_t input, const char* n, float volt = 0.0f);
+    Pin(uint8_t num, PinType t, uint8_t input, const char* n, uint16_t volt = 0);
 
     void setMode(uint8_t input);
     void setNote(const char* newNote);  // Metodo per cambiare la nota
